Separate non-Maven-specific code in OpCodeGen into a separate file
